Add command line usage output
Switch to Lucid - https://speakerdeck.com/patch/modern-getopt-for-command-line-processing-in-perl?slide=46

Better output if files or args missing.

Interesting edge cases - lucid accepts "bare" options, and these can be confused for filenames. Not an issue of the normal usage is followed.

This took longer than I planned to spend on it. No-one shocked.